+++ demos/randomCircle.js
@@ -5,7 +5,7 @@
 // +-------------------------------
 var randomCircle = function (viewport) {
   // var vp = pb.viewport();
-  var radius = 2 + Math.random() * viewport.getMinDimension() * 0.2;
+  var radius = 10 + Math.random() * viewport.getMinDimension() * 0.2;
   var center = viewport.randomPoint(radius, radius);
   var circle = new Circle(center, radius);
   return circle;