+++ src/esm/Circle.js
@@ -57,6 +57,20 @@ export class Circle {
         this.radius = radius;
     }
     /**
+     * Set the radius of this circle.
+     * The method is meant for chaining, you may also alter the `radius` attribute directly.
+     *
+     * @method setRadius
+     * @param {number} radius - The amount to move.
+     * @instance
+     * @memberof Circle
+     * @return {Circle} this for chaining
+     **/
+    setRadius(radius) {
+        this.radius = radius;
+        return this;
+    }
+    /**
      * Move the circle by the given amount.
      *
      * @method move